requirements.txt
azure-storage-blob==12.8.1
paramiko==3.3.1
pymsteams==0.2.2
PyYAML==6.0.1
setuptools==68.2.2
SQLAlchemy==2.0.25
pyodbc >= 5.0.0